+++ SpellsPlugin/src/com/elmakers/mine/bukkit/plugins/spells/builtin/UndoSpell.java
@@ -1,23 +1,51 @@
 package com.elmakers.mine.bukkit.plugins.spells.builtin;
 
 import org.bukkit.Material;
+import org.bukkit.block.Block;
 
-import com.elmakers.mine.bukkit.plugins.groups.PlayerPermissions;
 import com.elmakers.mine.bukkit.plugins.spells.Spell;
 
 public class UndoSpell extends Spell
 {
-
+	public UndoSpell()
+	{
+		addVariant("revert", Material.REDSTONE, getCategory(), "Deconstruct your target", "target");
+	}
+	
 	@Override
 	public boolean onCast(String[] parameters)
 	{
 		String undoPlayer = player.getName();
-		PlayerPermissions permissions = plugin.getPermissions(player.getName());
-		if (parameters.length > 0 && permissions.isAdministrator())
+		
+		for (int i = 0; i < parameters.length; i++)
 		{
-			undoPlayer = parameters[0];
+			if (parameters[i].equalsIgnoreCase("player") && i < parameters.length - 1)
+			{
+				undoPlayer = parameters[++i];
+			}
+			else if (parameters[i].equalsIgnoreCase("target"))
+			{
+				targetThrough(Material.GLASS);
+				Block target = getTargetBlock();
+				if (target == null)
+				{
+					castMessage(player, "No target");
+					return false;
+				}
+				
+				boolean undone = plugin.undo(undoPlayer, target);
+				if (undone)
+				{
+					castMessage(player, "You revert your construction");
+				}
+				else
+				{
+					castMessage(player, "That is not your construction");
+				}
+				return undone;
+			}
 		}
-			
+		
 		boolean undone = plugin.undo(undoPlayer);
 		if (undone)
 		{